% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverser.R
\name{reverser}
\alias{reverser}
\title{Reverses a string or number}
\usage{
reverser(toReverse)
}
\arguments{
\item{toReverse}{A String or number}
}
\value{
the reverse of the provided string or number
}
\description{
Reverses a string or number
}
\examples{
reverser("testit")
}
